Aplicacao/requirements.txt
Flask==2.1.3
Flask-Cors==3.0.10
flask-openapi3==2.1.0
Flask-SQLAlchemy==2.5.1
nose2==0.12.0
pydantic==1.10.2
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
typing_extensions==4.3.0
werkzeug==2.0
bcrypt
PyJWT
jsonify